remove base class -> seperate independent SDFT/SDPO
